Calculate and display date of specific pregnancy week

Calculate and display date of specific pregnancy week

Usage

date_when(weeks, due_date = NULL, person = NULL, today = Sys.Date())

Arguments

weeks

Numeric value indicating the number of weeks of pregnancy to calculate the date for

due_date

Date. The expected due date. If NULL, will try to use the "pregnancy.due_date" option. Required if option not set.

person

The person who is pregnant, to determine the grammar for the output message. Can be:

"I", "1", "1st", "first", or numeric 1 for first person

"you", "2", "2nd", "second", or numeric 2 for second person

Any other name for third person

NULL: will try to use the "pregnancy.person" option. Defaults to "You" if the option is set.

today

Date object representing the reference date for calculations. Default is Sys.Date(). This parameter exists primarily for testing and documentation purposes and it is unlikely to make sense for the user to need or want to change it from the default.

Value

Invisibly returns NULL. Prints messages to the console showing:

When the specified week of pregnancy occurs/occurred/will occur

How far in the past/future that date is from today (unless today is the date)

Details

The function calculates when someone will be/was a specific number of weeks pregnant based on their due date. It handles past, present and future dates appropriately in its messaging. The due date can be provided directly or set globally using options("pregnancy.due_date"). Similarly, the person being referenced can be provided directly or set globally using options("pregnancy.person").

See also

Examples

# Set a due date
due <- as.Date("2024-09-01")

# When will they be 12 weeks pregnant?
date_when(12, due_date = due)
#> ℹ On February 18, 2024, you were 12 weeks pregnant.
#> ℹ That was 51 weeks and 3 days ago.

# When will they be 20 weeks pregnant?
date_when(20, due_date = due, person = "Sarah")
#> ℹ On April 14, 2024, Sarah was 20 weeks pregnant.
#> ℹ That was 43 weeks and 3 days ago.