An example medications table

A data frame with example medications that might be used during fertility treatment/first trimester. It is an example of a data frame that might be used as the meds argument to meds_remaining().

Usage

medications

Format

A data frame with 14 rows and 5 columns:

medication

Name of the medication

format

Format of medication

quantity

Number taken per day

start_date

Date to start taking the medication

stop_date

Final date on which the medication is taken. See details.

Details

Note that the same medication (prednisolone in this example) has several rows, first because the quantity taken per day changes, then because it needs to be taken on non-consecutive days.

See also

Examples

medications
#> # A tibble: 14 × 5
#>    medication   format    quantity start_date stop_date 
#>    <chr>        <chr>        <dbl> <date>     <date>    
#>  1 progynova    pills            3 2025-04-21 2025-04-30
#>  2 LDA          pills            1 2025-04-21 2025-05-07
#>  3 prednisolone pills            2 2025-04-26 2025-05-07
#>  4 progynova    pills            6 2025-05-01 2025-07-11
#>  5 cyclogest    pessary          2 2025-05-03 2025-07-11
#>  6 lubion       injection        1 2025-05-03 2025-07-11
#>  7 clexane      injection        1 2025-05-08 2025-09-05
#>  8 prednisolone pills            4 2025-05-08 2025-07-11
#>  9 prednisolone pills            3 2025-07-12 2025-07-14
#> 10 prednisolone pills            2 2025-07-15 2025-07-17
#> 11 prednisolone pills            1 2025-07-18 2025-07-20
#> 12 prednisolone pills            1 2025-07-22 2025-07-22
#> 13 prednisolone pills            1 2025-07-24 2025-07-24
#> 14 prednisolone pills            1 2025-07-26 2025-07-26